button functionality
When you click on the Follow button, the text changes to Following. The color of the button also changes. And yours is added to the number of followers. That is, the initial number is 100,500 followers. When you click on the button, it will be 100,501.
When the page is updated, the final result of the user's actions should be recorded. That is, if you click on the button and refresh the page, the button still remains in the Following state with the appropriate color, and the number of followers does NOT decrease to the initial value.
When you click the button again, its text and color change to their original state. The number of followers also changes. It decreases by 1 (100,500).
In the code, the number 100,500 must be written with one value (100500). In the UI, it is displayed with a comma (100,500).